requirements.txt
# FastAPI backend
fastapi
uvicorn[standard]

# PDF extraction
pdfplumber

# Multi-language detection
langdetect
regex

# Date parsing
dateparser

# Pydantic (FastAPI uses pydantic v2)
pydantic>=2.0

# AI (Gemini)
google-generativeai

# Streamlit UI
streamlit
requests

# File uploads for FastAPI
python-multipart

# Optional (recommended) for environment variable loading
python-dotenv
